+++ src/Symfony/Bundle/WebServerBundle/WebServerConfig.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Bundle\WebServerBundle;
 
+use Symfony\Component\Process\PhpExecutableFinder;
+
 /**
  * @author Fabien Potencier <fabien@symfony.com>
  */
@@ -21,6 +23,7 @@ class WebServerConfig
     private $documentRoot;
     private $env;
     private $router;
+    private $executable;
 
     public function __construct(string $documentRoot, string $env, string $address = null, string $router = null)
     {
@@ -69,6 +72,22 @@ class WebServerConfig
         if (!ctype_digit($this->port)) {
             throw new \InvalidArgumentException(sprintf('Port "%s" is not valid.', $this->port));
         }
+
+        $executable = null;
+
+        $envExecutable = getenv('SYMFONY_SERVER_EXECUTABLE');
+        if ($envExecutable !== false) {
+            $executable = $envExecutable;
+        }
+
+        if ($executable === null) {
+            $finder = new PhpExecutableFinder();
+            if (false === $executable = $finder->find()) {
+                throw new \RuntimeException('Unable to find the PHP executable.');
+            }
+        }
+
+        $this->executable = $executable;
     }
 
     public function getDocumentRoot()
@@ -84,6 +103,11 @@ class WebServerConfig
     public function getRouter()
     {
         return $this->router;
+    }
+
+    public function getExecutable()
+    {
+        return $this->executable;
     }
 
     public function getHostname()