+++ src/Symfony/Bundle/WebServerBundle/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 -----
 
  * WebServer can now use '*' as a wildcard to bind to 0.0.0.0 (INADDR_ANY)
+ * Custom executable can be set using `SYMFONY_SERVER_EXECUTABLE` environment variable
 
 3.3.0
 -----